Move event tests to long suite.
'should emit hostAdd hostRemove' takes 2 minutes locally and even longer
in a constrained CI environment.
